refactor(server): split websocket boundary tests
Move the first low-risk websocket test clusters out of the monolithic tests.rs file and into child modules for dispatch/replay, registration, and websocket send helpers. The split keeps shared fixtures in the parent test module and does not touch production protocol code.

Constraint: This #350 slice is test organization only; XMPP/XEP behavior, namespaces, and wire shapes must remain unchanged.

Rejected: Big-bang split of all websocket tests | higher review risk and more import churn.

Confidence: high

Scope-risk: narrow

Directive: Continue splitting tests by behavior boundary; avoid moving shared fixtures into production modules.

Tested: cargo test -p waddle-server registration::

Tested: cargo test -p waddle-server send::

Tested: cargo test -p waddle-server dispatch::

Tested: cargo test -p waddle-server server::routes::websocket

Tested: cargo clippy -p waddle-server --tests -- -D warnings

Tested: git diff --check

Related: #350
