+++ server/crates/waddle-server/src/server/routes/websocket/tests.rs
@@ -3,12 +3,9 @@ use super::{
     cleanup::cleanup_connection_shutdown,
     frame::handle_xmpp_frame,
     interpret_loop::build_interpret_deps,
-    registration::{register_bound_connection_after_frame, RegistrationAfterFrame},
     replay::{drain_outbound_into_replay, drive_interpret_loop},
-    send::{send_ws_message, send_ws_text_frames},
-    session_init::load_blocklist_for_bind,
     state::WsConnState,
-    stream_management::{is_countable_stanza, SmRegistrationFinalization},
+    stream_management::is_countable_stanza,
     transport_xml::{build_stream_features_xml, sasl_failure_xml, sasl_success_xml},
 };
 use crate::config::ServerConfig;
@@ -16,13 +13,10 @@ use crate::db::{DatabaseConfig, DatabasePool, MigrationRunner, PoolConfig};
 use crate::permissions::{Object, ObjectType, Permission, Relation, Subject, Tuple, WriteTuple};
 use crate::server::bootstrap_membership::DEPLOYMENT_SERVER_ID;
 use crate::server::AppState;
-use futures::Sink;
 use hmac::{Hmac, Mac};
 use pbkdf2::pbkdf2_hmac;
 use sha2::{Digest, Sha256};
-use std::pin::Pin;
 use std::sync::Arc;
-use std::task::{Context, Poll};
 use tokio::sync::mpsc;
 // Handler functions moved to sub-modules but called directly in tests
 use handlers::iq::{
@@ -40,6 +34,10 @@ use waddle_xmpp::Affiliation;
 use xmpp_parsers::iq::{Iq, IqType};
 use xmpp_parsers::message::MessageType as XmppMessageType;
 
+mod dispatch;
+mod registration;
+mod send;
+
 #[test]
 fn service_domains_use_xmpp_domain_for_all_components() {
     let domains = XmppServiceDomains::new("waddle.social");
@@ -48,50 +46,6 @@ fn service_domains_use_xmpp_domain_for_all_components() {
     assert_eq!(domains.muc, "muc.waddle.social");
     assert_eq!(domains.spaces, "spaces.waddle.social");
     assert_eq!(domains.upload, "upload.waddle.social");
-}
-
-#[derive(Default)]
-struct TestSink {
-    fail_after: Option<usize>,
-    sent: Vec<Message>,
-}
-
-impl TestSink {
-    fn succeeds() -> Self {
-        Self::default()
-    }
-
-    fn fails_after(sent_before_failure: usize) -> Self {
-        Self {
-            fail_after: Some(sent_before_failure),
-            sent: Vec::new(),
-        }
-    }
-}
-
-impl Sink<Message> for TestSink {
-    type Error = &'static str;
-
-    fn poll_ready(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Poll::Ready(Ok(()))
-    }
-
-    fn start_send(mut self: Pin<&mut Self>, item: Message) -> Result<(), Self::Error> {
-        if matches!(self.fail_after, Some(limit) if self.sent.len() >= limit) {
-            return Err("synthetic websocket sink failure");
-        }
-
-        self.sent.push(item);
-        Ok(())
-    }
-
-    fn poll_flush(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Poll::Ready(Ok(()))
-    }
-
-    fn poll_close(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-        Poll::Ready(Ok(()))
-    }
 }
 
 async fn create_test_websocket_state() -> Arc<WebSocketState> {
@@ -564,36 +518,6 @@ fn authenticated_phase_for_session(session: &Session, domain: &str) -> Connectio
 }
 
 #[tokio::test]
-async fn send_ws_text_frames_stops_after_first_send_failure() {
-    let mut sink = TestSink::fails_after(1);
-
-    let sent = send_ws_text_frames(
-        &mut sink,
-        vec!["<open/>".to_string(), "<features/>".to_string()],
-        "synthetic failure",
-    )
-    .await;
-
-    assert!(!sent);
-    assert_eq!(sink.sent.len(), 1);
-}
-
-#[tokio::test]
-async fn send_ws_message_returns_true_on_success() {
-    let mut sink = TestSink::succeeds();
-
-    let sent = send_ws_message(
-        &mut sink,
-        Message::Text("<pong/>".into()),
-        "unexpected failure",
-    )
-    .await;
-
-    assert!(sent);
-    assert_eq!(sink.sent.len(), 1);
-}
-
-#[tokio::test]
 async fn handle_xmpp_frame_open_dispatches_via_typed_ingress() {
     let state = create_test_websocket_state().await;
     let mut conn = WsConnState::new();
@@ -646,590 +570,6 @@ async fn handle_xmpp_frame_malformed_auth_returns_malformed_request() {
     .await;
 
     assert_eq!(responses, vec![sasl_failure_xml("malformed-request")]);
-}
-
-// ---------------------------------------------------------------
-// #229 PR11 — DeliveryKind dispatch in the per-connection main loop
-// ---------------------------------------------------------------
-//
-// The actual main-loop entry point is `xmpp_websocket_handler`, an
-// async function tied to a real WebSocket sink. To test the
-// dispatch logic in isolation we exercise its two helpers
-// (`build_interpret_deps`, `drive_interpret_loop`) and the
-// `WsConnState::ensure_state_machine` lifecycle directly. End-to-
-// end coverage of the routing flow lands once PR12 emits
-// `OutboundStanza::peer_stanza` from `RouteToConnection`.
-
-#[tokio::test]
-async fn ensure_state_machine_initializes_sm_in_ready_phase() {
-    let state = create_test_websocket_state().await;
-    let mut conn = WsConnState::new();
-    let jid: jid::FullJid = "alice@example.com/web".parse().expect("jid");
-
-    assert!(
-        conn.state_machine.is_none(),
-        "fresh WsConnState has no state machine"
-    );
-
-    conn.ensure_state_machine(
-        "example.com",
-        &state.deps.protocol.dispatcher,
-        jid.clone(),
-        false,
-        Blocklist::empty(),
-    );
-
-    let sm = conn.state_machine.as_ref().expect("SM initialized");
-    assert!(matches!(sm.phase(), ConnectionPhase::Ready { .. }));
-    assert_eq!(sm.phase().bound_jid(), Some(&jid));
-}
-
-#[tokio::test]
-async fn register_bound_connection_after_frame_registers_ready_connection_once() {
-    let state = create_test_websocket_state().await;
-    let mut conn = WsConnState::new();
-    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
-    let (tx, _rx) = mpsc::channel::<OutboundStanza>(1);
-    let mut pending_tx = Some(tx);
-
-    conn.phase = ConnectionPhase::ready(jid.clone(), false);
-    conn.carbons_enabled = true;
-    conn.roster_interested = true;
-    conn.presence_available = true;
-    conn.presence_show = Some(xmpp_parsers::presence::Show::Chat);
-    conn.presence_status = Some("ready".to_string());
-    conn.presence_priority = 7;
-
-    let result = register_bound_connection_after_frame(
-        state.as_ref(),
-        "example.com",
-        &mut conn,
-        &mut pending_tx,
-    )
-    .await;
-
-    assert!(matches!(
-        result,
-        RegistrationAfterFrame::Registered(SmRegistrationFinalization::KeepExistingResponses)
-    ));
-    assert!(
-        pending_tx.is_none(),
-        "registration consumes the one-shot sender"
-    );
-    assert!(
-        conn.registry_owner.is_some(),
-        "registry ownership is tracked"
-    );
-
-    let sm = conn
-        .state_machine
-        .as_ref()
-        .expect("state machine initialized");
-    assert!(matches!(sm.phase(), ConnectionPhase::Ready { .. }));
-    assert_eq!(sm.phase().bound_jid(), Some(&jid));
-
-    let entry = state
-        .deps
-        .protocol
-        .connection_registry
-        .get_entry(&jid)
-        .expect("registered connection");
-    assert!(entry.is_carbons_enabled());
-    assert!(entry
-        .roster_interested
-        .load(std::sync::atomic::Ordering::Relaxed));
-    assert!(entry.is_presence_available());
-    assert_eq!(entry.presence_priority(), 7);
-
-    let presence = state
-        .deps
-        .protocol
-        .connection_registry
-        .get_presence_state(&jid)
-        .expect("presence state restored");
-    assert_eq!(presence.show.as_deref(), Some("chat"));
-    assert_eq!(presence.status.as_deref(), Some("ready"));
-    assert_eq!(presence.priority, 7);
-
-    let second = register_bound_connection_after_frame(
-        state.as_ref(),
-        "example.com",
-        &mut conn,
-        &mut pending_tx,
-    )
-    .await;
-    assert!(matches!(second, RegistrationAfterFrame::Unchanged));
-    assert_eq!(
-        state.deps.protocol.connection_registry.connection_count(),
-        1
-    );
-}
-
-#[tokio::test]
-async fn register_bound_connection_after_frame_completes_pending_resume_claim() {
-    use waddle_xmpp::stream_management::{
-        DetachedSession, DetachedUnackedStanza, SmSessionRegistry,
-    };
-
-    let state = create_test_websocket_state().await;
-    let mut conn = WsConnState::new();
-    let jid: FullJid = "alice@example.com/web".parse().expect("jid");
-    let stream_id = "registration-resume-stream".to_string();
-    let session = create_test_session(state.as_ref(), "alice").await;
-
-    state
-        .deps
-        .protocol
-        .resumable_sessions
-        .insert(stream_id.clone(), session.clone());
-    state
-        .deps
-        .protocol
-        .sm_session_registry
-        .store_session(DetachedSession {
-            stream_id: stream_id.clone(),
-            user_id: session.user_id.clone(),
-            jid: jid.clone(),
-            inbound_count: 4,
-            outbound_count: 10,
-            last_acked: 8,
-            replay_gap_through: None,
-            unacked_stanzas: vec![
-                DetachedUnackedStanza {
-                    sequence: 9,
-                    stanza_xml: "<message id='m9'/>".to_string(),
-                    original_receipt_at: chrono::Utc::now(),
-                },
-                DetachedUnackedStanza {
-                    sequence: 10,
-                    stanza_xml: "<message id='m10'/>".to_string(),
-                    original_receipt_at: chrono::Utc::now(),
-                },
-            ],
-            max_resume_time: Some(300),
-            detached_at: std::time::Instant::now(),
-            carbons_enabled: true,
-            roster_interested: true,
-            presence_available: true,
-            presence_show: Some(xmpp_parsers::presence::Show::Chat),
-            presence_status: Some("back".to_string()),
-            presence_priority: 5,
-        })
-        .await
-        .expect("store detached session");
-
-    conn.phase = ConnectionPhase::authenticated(&jid);
-    conn.authenticated_session = Some(session.clone());
-    let resume_frame = format!("<resume xmlns='urn:xmpp:sm:3' previd='{stream_id}' h='9'/>");
-    let resume_responses =
-        handle_xmpp_frame(&resume_frame, "example.com", state.as_ref(), &mut conn).await;
-
-    assert!(!resume_responses.is_empty());
-    assert_eq!(
-        conn.pending_resume_stream_id.as_deref(),
-        Some(stream_id.as_str())
-    );
-    assert_eq!(conn.pending_resume_h, Some(9));
-    assert!(conn.suppress_sm_record_next_batch);
-
-    let (tx, _rx) = mpsc::channel::<OutboundStanza>(1);
-    let mut pending_tx = Some(tx);
-    let result = register_bound_connection_after_frame(
-        state.as_ref(),
-        "example.com",
-        &mut conn,
-        &mut pending_tx,
-    )
-    .await;
-
-    match result {
-        RegistrationAfterFrame::Registered(SmRegistrationFinalization::ReplaceWithResumed {
-            resumed,
-            replay_after_h,
-        }) => {
-            assert_eq!(resumed.previd.as_str(), stream_id.as_str());
-            assert_eq!(resumed.h, 4);
-            assert_eq!(replay_after_h, 9);
-        }
-        _ => panic!("expected resumed finalization"),
-    }
-
-    assert!(
-        pending_tx.is_none(),
-        "registration consumes the one-shot sender"
-    );
-    assert!(conn.pending_resume_stream_id.is_none());
-    assert!(conn.pending_resume_h.is_none());
-    assert_eq!(
-        conn.sm_state.get_stanzas_to_resend(9),
-        vec!["<message id='m10'/>".to_string()]
-    );
-    assert!(state
-        .deps
-        .protocol
-        .resumable_sessions
-        .get(&stream_id)
-        .is_none());
-    assert!(state
-        .deps
-        .protocol
-        .sm_session_registry
-        .peek_session(&stream_id)
-        .await
-        .expect("peek detached session")
-        .is_none());
-
-    let entry = state
-        .deps
-        .protocol
-        .connection_registry
-        .get_entry(&jid)
-        .expect("registered resumed connection");
-    assert!(entry.is_carbons_enabled());
-    assert!(entry
-        .roster_interested
-        .load(std::sync::atomic::Ordering::Relaxed));
-    assert!(entry.is_presence_available());
-    assert_eq!(entry.presence_priority(), 5);
-
-    let presence = state
-        .deps
-        .protocol
-        .connection_registry
-        .get_presence_state(&jid)
-        .expect("resumed presence state restored");
-    assert_eq!(presence.show.as_deref(), Some("chat"));
-    assert_eq!(presence.status.as_deref(), Some("back"));
-    assert_eq!(presence.priority, 5);
-}
-
-#[tokio::test]
-async fn ensure_state_machine_seeds_blocklist_from_database_at_bind() {
-    // #229 PR13: bind-time SM seeding from
-    // `DatabaseBlockingStorage`. Persist a single blocked entry
-    // for alice, run the bind-time loader against the same
-    // global pool, hand the result to `ensure_state_machine`,
-    // then drive a synchronous dispatch through a probe handler
-    // and observe the seeded entry on the `MessageContext`
-    // snapshot. Without the seed, the snapshot would be
-    // `Blocklist::empty()` and `BlockingFilterHandler` (post
-    // PR16 cutover) would silently regress XEP-0191 enforcement.
-    use crate::db::blocking::DatabaseBlockingStorage;
-    use std::sync::Mutex;
-    use waddle_xmpp::protocol::{HandlerOutcome, MessageContext, MessageHandler};
-
-    let state = create_test_websocket_state().await;
-    let alice_full: jid::FullJid = "alice@example.com/web".parse().expect("jid");
-    let alice_bare = alice_full.to_bare();
-    let blocked_bare: BareJid = "blocked@example.com".parse().expect("bare");
-
-    // Seed persistence with one entry.
-    let storage = DatabaseBlockingStorage::new(state.deps.app_state.db_pool.global().clone());
-    storage
-        .add_blocks(&alice_bare, &[blocked_bare.to_string()])
-        .await
-        .expect("add_blocks");
-
-    // Mirror the bind-site loader.
-    let blocklist = load_blocklist_for_bind(&state.deps.app_state.db_pool, &alice_full)
-        .await
-        .expect("blocklist load succeeds when storage is healthy");
-    let loaded: Vec<_> = blocklist.iter().cloned().collect();
-    assert_eq!(loaded, vec![blocked_bare.clone()]);
-
-    // Build a probe-only dispatcher so the assertion isolates
-    // the SM seeding behaviour from any side effects of the
-    // production message-pipeline chain (those have their own
-    // dedicated tests). The goal here is "the seeded blocklist
-    // shows up on the `MessageContext` snapshot".
-    let captured: Arc<Mutex<Vec<waddle_xmpp::protocol::Blocklist>>> =
-        Arc::new(Mutex::new(Vec::new()));
-    struct SnapshotProbe {
-        captured: Arc<Mutex<Vec<waddle_xmpp::protocol::Blocklist>>>,
-    }
-    impl MessageHandler for SnapshotProbe {
-        fn name(&self) -> &'static str {
-            "ws-bind-blocklist-probe"
-        }
-        fn handle(
-            &self,
-            _message: &mut xmpp_parsers::message::Message,
-            ctx: &MessageContext<'_>,
-        ) -> HandlerOutcome {
-            self.captured
-                .lock()
-                .expect("mutex")
-                .push(ctx.blocklist.clone());
-            HandlerOutcome::Continue(Vec::new())
-        }
-    }
-    let mut probe_dispatcher = StanzaDispatcher::new();
-    probe_dispatcher.register_message(Arc::new(SnapshotProbe {
-        captured: captured.clone(),
-    }));
-    let dispatcher = Arc::new(probe_dispatcher);
-
-    let mut conn = WsConnState::new();
-    conn.ensure_state_machine(
-        "example.com",
-        &dispatcher,
-        alice_full.clone(),
-        false,
-        blocklist,
-    );
-
-    // Drive a chat message dispatch so the probe fires.
-    let mut msg =
-        xmpp_parsers::message::Message::new(Some("bob@example.com".parse().expect("to jid")));
-    msg.from = Some(jid::Jid::from(alice_full.clone()));
-    msg.type_ = XmppMessageType::Chat;
-    msg.bodies.insert(
-        String::new(),
-        xmpp_parsers::message::Body("hello".to_string()),
-    );
-    let sm = conn.state_machine.as_mut().expect("SM");
-    sm.handle(InboundEvent::FrameReceived(InboundFrame::Stanza(Box::new(
-        Stanza::Message(msg),
-    ))));
-
-    let snapshots = captured.lock().expect("mutex").clone();
-    assert_eq!(snapshots.len(), 1, "probe runs exactly once");
-    let entries: Vec<_> = snapshots[0].iter().cloned().collect();
-    assert_eq!(
-        entries,
-        vec![blocked_bare],
-        "MessageContext snapshot must reflect the persisted blocklist"
-    );
-}
-
-#[tokio::test]
-async fn drive_interpret_loop_resolves_send_stanza_into_wire_frames() {
-    // Recipient pass produces `OutboundEvent::SendStanza` for the
-    // wire write. Drive the loop with a single SendStanza event
-    // and assert it serializes cleanly into a frame (no extra
-    // round-trips through the SM since no callback feedback is
-    // produced).
-    let state = create_test_websocket_state().await;
-    let mut conn = WsConnState::new();
-    let jid: jid::FullJid = "bob@example.com/desk".parse().expect("jid");
-    conn.ensure_state_machine(
-        "example.com",
-        &state.deps.protocol.dispatcher,
-        jid,
-        false,
-        Blocklist::empty(),
-    );
-    let sm = conn.state_machine.as_mut().expect("SM");
-
-    let mut msg =
-        xmpp_parsers::message::Message::new(Some("alice@example.com".parse().expect("to jid")));
-    msg.from = Some("bob@example.com/desk".parse().expect("from jid"));
-    msg.type_ = xmpp_parsers::message::MessageType::Chat;
-    msg.bodies.insert(
-        String::new(),
-        xmpp_parsers::message::Body("hello".to_string()),
-    );
-
-    let initial_events = vec![OutboundEvent::SendStanza(Box::new(Stanza::Message(msg)))];
-    let deps = build_interpret_deps(state.as_ref(), None);
-    let (frames, close) = drive_interpret_loop(initial_events, sm, &deps).await;
-
-    assert!(!close, "SendStanza alone never requests transport close");
-    assert_eq!(frames.len(), 1, "single SendStanza → single wire frame");
-    assert!(
-        frames[0].contains("hello"),
-        "wire frame carries the message body; got {:?}",
-        frames[0]
-    );
-}
-
-#[tokio::test]
-async fn drive_interpret_loop_runs_recipient_pass_for_peer_message() {
-    // Production-shape regression: feed `InboundEvent::StanzaFromPeer`
-    // through a Ready state machine and drive the resulting events
-    // via `drive_interpret_loop`. The recipient pass MUST produce
-    // a wire frame containing bob's recipient-side `<stanza-id>`
-    // stamp so XEP-0359 §5 conformance is preserved end-to-end
-    // through the production helpers.
-    let state = create_test_websocket_state().await;
-    let mut conn = WsConnState::new();
-    let bob_full: jid::FullJid = "bob@example.com/desk".parse().expect("jid");
-    conn.ensure_state_machine(
-        "example.com",
-        &state.deps.protocol.dispatcher,
-        bob_full,
-        false,
-        Blocklist::empty(),
-    );
-    let sm = conn.state_machine.as_mut().expect("SM");
-
-    let mut peer_msg =
-        xmpp_parsers::message::Message::new(Some("bob@example.com".parse().expect("to jid")));
-    peer_msg.from = Some("alice@example.com/web".parse().expect("from jid"));
-    peer_msg.type_ = xmpp_parsers::message::MessageType::Chat;
-    peer_msg.id = Some("alice-wire-id".to_string());
-    peer_msg.bodies.insert(
-        String::new(),
-        xmpp_parsers::message::Body("hi bob".to_string()),
-    );
-    // Pre-stamp alice's sender-side stanza-id so we can verify
-    // the recipient pass *adds* bob's stamp rather than replacing
-    // alice's (XEP-0359 §5 cross-archive preservation).
-    peer_msg
-        .payloads
-        .push(waddle_xmpp_core::xep0359::build_stanza_id_element(
-            "alice-A1",
-            &"alice@example.com".parse::<jid::Jid>().expect("jid"),
-        ));
-
-    let events = sm.handle(InboundEvent::StanzaFromPeer(Box::new(Stanza::Message(
-        peer_msg,
-    ))));
-    let deps = build_interpret_deps(state.as_ref(), None);
-    let (frames, _close) = drive_interpret_loop(events, sm, &deps).await;
-
-    // Recipient pass terminates with at least one SendStanza
-    // carrying bob's stamp.
-    assert!(
-        !frames.is_empty(),
-        "recipient pass must produce at least one wire frame"
-    );
-    let combined = frames.join("\n");
-    assert!(
-        combined.contains("by=\"bob@example.com\""),
-        "recipient-pass wire frame must carry bob's stanza-id stamp; got: {combined}"
-    );
-    assert!(
-        combined.contains("alice-A1"),
-        "recipient-pass wire frame must preserve alice's cross-archive stanza-id; \
-             got: {combined}"
-    );
-    assert!(
-        combined.contains("hi bob"),
-        "recipient-pass wire frame must carry the message body; got: {combined}"
-    );
-}
-
-#[tokio::test]
-async fn drain_outbound_dispatches_direct_frame_into_unacked_unchanged() {
-    // Regression for the detach-drain DeliveryKind dispatch
-    // Qodo flagged on PR269: DirectFrame values must be recorded
-    // byte-for-byte (no recipient pipeline). This is the live
-    // contract the SM-resume replay path depends on.
-    let state = create_test_websocket_state().await;
-    let mut conn = WsConnState::new();
-    let jid: jid::FullJid = "bob@example.com/desk".parse().expect("jid");
-    conn.ensure_state_machine(
-        "example.com",
-        &state.deps.protocol.dispatcher,
-        jid,
-        false,
-        Blocklist::empty(),
-    );
-    // Enable SM tracking so `record_outbound` actually retains the
-    // drained XML.
-    conn.sm_state.enabled = true;
-
-    let mut msg =
-        xmpp_parsers::message::Message::new(Some("alice@example.com".parse().expect("to jid")));
-    msg.from = Some("bob@example.com/desk".parse().expect("from jid"));
-    msg.type_ = xmpp_parsers::message::MessageType::Chat;
-    msg.bodies.insert(
-        String::new(),
-        xmpp_parsers::message::Body("plain".to_string()),
-    );
-    let expected_xml = stanza_to_xml(&Stanza::Message(msg.clone()));
-
-    let (tx, mut rx) = mpsc::channel::<OutboundStanza>(4);
-    tx.send(OutboundStanza::new(Stanza::Message(msg)))
-        .await
-        .expect("send");
-    drop(tx); // close so try_recv eventually returns Empty
-
-    drain_outbound_into_replay(
-        state.as_ref(),
-        conn.state_machine.as_mut(),
-        &mut conn.sm_state,
-        None,
-        &mut rx,
-        None,
-    )
-    .await;
-
-    let queue = conn.sm_state.get_stanzas_to_resend(0);
-    assert_eq!(queue.len(), 1, "DirectFrame recorded once");
-    assert_eq!(
-        queue[0], expected_xml,
-        "DirectFrame is recorded byte-for-byte (no recipient pipeline rewrite)"
-    );
-}
-
-#[tokio::test]
-async fn drain_outbound_dispatches_peer_stanza_through_recipient_pass() {
-    // PeerStanza values queued during detach must run through
-    // the recipient pass before being recorded in the SM unacked
-    // queue, so a resumed connection's replay carries the
-    // recipient-side `<stanza-id>` stamp. Without the dispatch
-    // (Qodo's flagged bug), the queued bytes would be the raw
-    // peer stanza missing bob's stamp.
-    let state = create_test_websocket_state().await;
-    let mut conn = WsConnState::new();
-    let bob_full: jid::FullJid = "bob@example.com/desk".parse().expect("jid");
-    conn.ensure_state_machine(
-        "example.com",
-        &state.deps.protocol.dispatcher,
-        bob_full,
-        false,
-        Blocklist::empty(),
-    );
-    conn.sm_state.enabled = true;
-
-    let mut peer_msg =
-        xmpp_parsers::message::Message::new(Some("bob@example.com".parse().expect("to jid")));
-    peer_msg.from = Some("alice@example.com/web".parse().expect("from jid"));
-    peer_msg.type_ = xmpp_parsers::message::MessageType::Chat;
-    peer_msg.id = Some("alice-wire-id".to_string());
-    peer_msg.bodies.insert(
-        String::new(),
-        xmpp_parsers::message::Body("hi from drain".to_string()),
-    );
-    peer_msg
-        .payloads
-        .push(waddle_xmpp_core::xep0359::build_stanza_id_element(
-            "alice-A1",
-            &"alice@example.com".parse::<jid::Jid>().expect("jid"),
-        ));
-
-    let (tx, mut rx) = mpsc::channel::<OutboundStanza>(4);
-    tx.send(OutboundStanza::peer_stanza(Stanza::Message(peer_msg)))
-        .await
-        .expect("send");
-    drop(tx);
-
-    drain_outbound_into_replay(
-        state.as_ref(),
-        conn.state_machine.as_mut(),
-        &mut conn.sm_state,
-        None,
-        &mut rx,
-        None,
-    )
-    .await;
-
-    let queue = conn.sm_state.get_stanzas_to_resend(0);
-    assert!(
-        !queue.is_empty(),
-        "PeerStanza drain MUST record at least the recipient-pass wire frame"
-    );
-    let combined: String = queue.join("\n");
-    assert!(
-        combined.contains("by=\"bob@example.com\""),
-        "drained PeerStanza replay must carry bob's recipient-side stanza-id; got: {combined}"
-    );
-    assert!(
-        combined.contains("alice-A1"),
-        "drained PeerStanza replay must preserve alice's cross-archive stamp; got: {combined}"
-    );
 }
 
 #[tokio::test]
