+++ server/crates/waddle-server/src/server/routes/websocket/tests/send.rs
@@ -1,0 +1,79 @@
+use super::super::send::{send_ws_message, send_ws_text_frames};
+use axum::extract::ws::Message;
+use futures::Sink;
+use std::pin::Pin;
+use std::task::{Context, Poll};
+
+#[derive(Default)]
+struct TestSink {
+    fail_after: Option<usize>,
+    sent: Vec<Message>,
+}
+
+impl TestSink {
+    fn succeeds() -> Self {
+        Self::default()
+    }
+
+    fn fails_after(sent_before_failure: usize) -> Self {
+        Self {
+            fail_after: Some(sent_before_failure),
+            sent: Vec::new(),
+        }
+    }
+}
+
+impl Sink<Message> for TestSink {
+    type Error = &'static str;
+
+    fn poll_ready(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
+    }
+
+    fn start_send(mut self: Pin<&mut Self>, item: Message) -> Result<(), Self::Error> {
+        if matches!(self.fail_after, Some(limit) if self.sent.len() >= limit) {
+            return Err("synthetic websocket sink failure");
+        }
+
+        self.sent.push(item);
+        Ok(())
+    }
+
+    fn poll_flush(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
+    }
+
+    fn poll_close(self: Pin<&mut Self>, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
+        Poll::Ready(Ok(()))
+    }
+}
+
+#[tokio::test]
+async fn send_ws_text_frames_stops_after_first_send_failure() {
+    let mut sink = TestSink::fails_after(1);
+
+    let sent = send_ws_text_frames(
+        &mut sink,
+        vec!["<open/>".to_string(), "<features/>".to_string()],
+        "synthetic failure",
+    )
+    .await;
+
+    assert!(!sent);
+    assert_eq!(sink.sent.len(), 1);
+}
+
+#[tokio::test]
+async fn send_ws_message_returns_true_on_success() {
+    let mut sink = TestSink::succeeds();
+
+    let sent = send_ws_message(
+        &mut sink,
+        Message::Text("<pong/>".into()),
+        "unexpected failure",
+    )
+    .await;
+
+    assert!(sent);
+    assert_eq!(sink.sent.len(), 1);
+}
